Create Assignment_02(code2)
Implement a problem of maximize Profit by trading stocks based on given rate per day. Statement: Given an array arr[] of N positive integers which denotes the cost of selling and buying a stock on each of the N days. The task is to find the maximum profit that can be earned by buying a stock on or selling all previously bought stocks on a particular day. Input: arr[] = {2, 3, 5} Output: 5 Input: arr[] = {8, 5, 1} Output: 0